Made nnueMul 32 bit friendly and added a comment explaining its purpose

+++ src/nnue/util.c
@@ -3,8 +3,8 @@
 
 #include "util.h"
 
-int32_t nnueMul(int8_t a, int8_t b) {
-	return ((int32_t)a)*((int32_t)b);
+int32_t nnueMul(int32_t a, int32_t b) {
+	return a*b;
 }
 
 int32_t nnueCReLU(int32_t x) {
